+++ src/base/b-window/interface.ts
@@ -1,0 +1,13 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+export type TitleValue<T = unknown> = string | ((ctx: T) => string);
+
+export interface StageTitles<T = unknown> extends Dictionary<TitleValue<T>> {
+	'[[DEFAULT]]'?: TitleValue<T>;
+}