+++ CHANGELOG.md
@@ -34,6 +34,31 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * `traits/i-history`
   * Next calling of `initIndex` pushes another one index page
 
+## v3.0.0-rc.121 (2021-01-12)
+
+#### :boom: Breaking Change
+
+* `bWindow`:
+  * Renamed the slot `control` to `controls`
+  * Replaced the global modifier `hidden` to `opened`
+
+#### :bug: Bug Fix
+
+* Fixed a bug with `getRootMod` `iStaticPage`
+* Fixed a bug with parsing of styles `core/component/vnode`
+
+#### :memo: Documentation
+
+* Added documentation `bWindow`
+* Improved jsDoc `iBlock`
+
+#### :house: Internal
+
+`bWindow`:
+  * Fixed ESLint errors
+  * Fixed TS errors
+  * Added tests
+
 ## v3.0.0-rc.120 (2020-12-23)
 
 #### :bug: Bug Fix